ImprovedModel.py
Hi,
I tried 5 different models and chose the "bert-large-uncased-whole-word-masking-finetuned-squad" model as it gave the most accurate answer when the answers were descriptive. The results were the same and accurate when answers were of one word, but the main difference was eminent in subjective answers. Output prints the question along with the answers.