Easy-to-use UI for Non technical clients
Hey!!
Instead of inputting multiple questions at once in an excel sheet and finding the answers by running the python code, I created a web interface in which clients from a non technical background can input their question and context in the "Question" and "Paragraph" section of the webpage respectively. The answer would be displayed in the "Output" section.
After running the code here, you will be redirected to a new tab on your browser, post which, you can present it to the client to experiment with the question-answering model.